requirements.txt
# Core
Django>=4.2,<5.0
python-dotenv>=1.0,<2.0

# OCR e processamento de PDF
pdfplumber>=0.11,<1.0
pytesseract>=0.3,<1.0
Pillow>=10.0,<12.0

# Geração e manipulação de arquivos
reportlab>=4.0,<5.0

# Produção
gunicorn>=21.2,<24.0
whitenoise>=6.5,<7.0

# Desenvolvimento e testes
django-extensions>=3.2,<4.0
pytest>=7.4,<9.0
pytest-django>=4.5,<5.0
black>=23.7,<26.0
flake8>=6.0,<8.0
